+++ embedded/esp32/lib/module_handler/include/modules/wifi_userconfig.hpp
@@ -61,6 +61,11 @@ class ModuleUserConfig : public Module {
    * @brief Start the user config webserver
    */
   void start();
+
+  /**
+   * @brief Handle clear buffer command and prepare acknowledgement response
+   */
+  void clearBuffer();
 };
 
 }  // namespace ModuleHandler